Enzyme Jacobian for PhasorDynamics::Load within GridKit (#131)
* Separated out Enzyme macros from FindEnzyme and added a EnzymeAddLibrary macro.

* Enzyme wrapper within GridKit.

* Fixes to rebase.

* More interesting vector residual for Enzyme examples.

* Fix some warnings.

* Address one more warning.

* Apply pre-commmit fixes

* Empty commit to trigger CI after pre-commit fixes.

* Successful use of target_compiile_options to pass enzyme-auto-sparsity flags in examples.

* More flags needed to get the correct answer to machine precision in EnzymePowerElectronicsCheck.

* EnzymeAddLibrary not currently needed.

* Apply pre-commmit fixes

* Empty commit to trigger CI after pre-commit fixes.

* Updated documentation in EnzymeAddLibrary.

* Fix warnings previously hidden by manual builds through enzyme_add_executable CMake macro.

* Use GridKit::Testing::isEqual in Enzyme examples.

* Fixes to rebase.

* Apply pre-commmit fixes

* Empty commit to trigger CI after pre-commit fixes.

* residual_wrapper can now be placed within GridKit::Enzyme namespace.

---------

+++ examples/Enzyme/Library/Vector/VectorModel.hpp
@@ -17,7 +17,7 @@ private:
   void                square(std::vector<double>&, std::vector<double>&);
 
 public:
-  VectorModel(int);
+  VectorModel(size_t);
   void                 setVariable(std::vector<double>);
   void                 evalResidual();
   void                 evalJacobian();
